+++ aclpy/client/client_pkcs12.py
@@ -3,8 +3,11 @@
 """Arrowhead Client class using .p12 certificates.
 """
 
+from typing import List
+
 from aclpy.client.client import ArrowheadClient as ArrowheadClientBase
 from aclpy.connector.connector_pkcs12 import ArrowheadConnector
+from aclpy.interface import ArrowheadInterface
 from aclpy.server import ArrowheadServer
 
 
@@ -17,6 +20,7 @@ class ArrowheadClient(ArrowheadClientBase):
     pubfile (str) -- path to the public key .pub
     cafile (str) -- path to the certificate authority file .ca
     server (ArrowheadServer) -- configuration of the Arrowhead Core server
+    interfaces (List[ArrowheadInterfaces]) -- list of available interfaces, [] by default
     """
 
     def __init__(self, *,
@@ -27,7 +31,8 @@ class ArrowheadClient(ArrowheadClientBase):
             p12pass: str,
             pubfile: str,
             cafile: str,
-            server: ArrowheadServer
+            server: ArrowheadServer,
+            interfaces: List[ArrowheadInterface] = [],
     ):
         """Initialize ArrowheadClient class."""
         # Read pubkey first
@@ -36,9 +41,12 @@ class ArrowheadClient(ArrowheadClientBase):
 
         self.connector = ArrowheadConnector(server)
 
-        super(ArrowheadClient, self).__init__(name, address, port, "".join(pubkey.split("\n")[1:-2]), self.connector)
+        super(ArrowheadClient, self).__init__(name, address, port, pubkey.replace("\n", ""), self.connector)
 
         self.p12file = p12file
         self.p12pass = p12pass
         self.pubfile = pubfile
         self.cafile = cafile
+
+        for interface in interfaces:
+            self.interfaces.append(interface)